Move checking if a command is scheduled already to an abstract CommandScheduler base.
